Fix teleop script using refactored codebase (#48)
* fix things for teleop

* make a bunch of fixes to start teleop scritp

* update dependencies

* updare run controller setup

* update setup for running code on the real robot
