+++ src/home_robot/setup.py
@@ -4,7 +4,15 @@
 # LICENSE file in the root directory of this source tree.
 from setuptools import setup, find_packages
 
-install_requires = ["numpy<1.24", "scipy", "hydra-core", "yacs"]
+install_requires = [
+    "numpy<1.24",
+    "scipy",
+    "hydra-core",
+    "yacs",
+    "h5py",
+    "pybullet",
+    "pygifsicle",
+]
 
 setup(
     name="home-robot",
