+++ src/home_robot_hw/launch/controller.launch
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
 <launch>
 	<node name="joy_node" type="joy_node" pkg="joy"/>
-	<node name="run_controller" type="run_controller_teleop.py" pkg="home_robot_hw"/>
+	<node name="run_controller" type="run_controller_teleop.py" pkg="home_robot_hw" output="screen"/>
 </launch>
