+++ src/home_robot_hw/scripts/run_controller_teleop.py
@@ -8,7 +8,7 @@
 import rospy
 import os
 from home_robot_hw.teleop.stretch_xbox_controller import StretchXboxController
-from home_robot.agent.motion.robot import HelloStretch
+from home_robot.agent.motion.stretch import HelloStretch
 from home_robot_hw.ros.path import get_urdf_dir
 
 
@@ -16,7 +16,6 @@ if __name__ == "__main__":
     rospy.init_node("xbox_controller")
 
     stretch_planner_urdf_path = get_urdf_dir()
-    print(stretch_planner_urdf_path)
     model = HelloStretch(
         visualize=False,
         root="",
